requirements.txt
Appium-Python-Client==3.2.1
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
black==24.8.0
certifi==2024.7.4
click==8.1.7
Django==4.2.15
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
mypy-extensions==1.0.0
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
PySocks==1.7.1
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
